BaseFiles
This is files That working together.
Main file is "Электронная заявка на пропуск". Base is - "SystemEZP". Cheblons - contain Print templates. "Инструкции" - Contain Images for instructions, that used in the Programm. "Sounds" contain Sounds that use the programm. SendMail Contain files that uses for sending mail to Head of the Pass Office. Archive contain an archive database